+++ api/src/models/mediaManager/Directory.ts
@@ -1,7 +1,6 @@
 import {IDirectory} from '../../../../shared/models/mediaManager/IDirectory';
 import {File} from './File';
 import * as mongoose from 'mongoose';
-import {Unit} from '../units/Unit';
 
 interface IDirectoryModel extends IDirectory, mongoose.Document {
 
@@ -56,7 +55,12 @@ directorySchema.pre('remove', async function () {
         await model.remove();
       }
     }
-    await Unit.deleteMany({'_id': {$in: localDir.files}}).exec();
+    for (const file of localDir.files) {
+      const model = await File.findById(file);
+      if (model) {
+        await model.remove();
+      }
+    }
   } catch (err) {
     throw new Error('Delete Error: ' + err.toString());
   }